Add teaching module and update rspec tests
