+++ student.rb
@@ -1,15 +1,12 @@
-class Student
+require_relative 'teacher'
+
+class Student < Teacher
   attr_reader :age, :phase
   attr_accessor :name
 
   def initialize(options = {})
+    super
     @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-  end
-
-  def offer_high_five
-    "High five!"
   end
 
   def set_phase(num)
@@ -30,4 +27,15 @@ class Student
     response += "Now I feel like a genius!"
     response
   end
+
+  def teach_stuff
+    raise NoMethodError
+    return false
+  end
+
+  def receive_raise(raise)
+    raise NoMethodError
+    return false
+  end
+
 end
