+++ teacher.rb
@@ -1,7 +1,12 @@
+require_relative 'teaching_mod'
+
 class Teacher
   attr_reader :age, :salary, :phase, :performance_rating, :target_raise
   attr_accessor :name
+
   RATINGCONSTANT = 90
+  include Teaching
+
   def initialize(options={})
     @phase = 3
     @age = options.fetch(:age, 0)
@@ -14,18 +19,18 @@ class Teacher
     "High five!"
   end
 
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
+  # def set_phase(num)
+  #   @phase = num
+  #   "Cool, I've always wanted to teach phase #{num}!"
+  # end
 
-  def teach_stuff
-    response = ""
-    response += "Listen, class, this is how everything works, fo SHO! "
-    response += "*drops flat-out insane knowledge bomb* "
-    response += "... You're welcome. *saunters away*"
-    response
-  end
+  # def teach_stuff
+  #   response = ""
+  #   response += "Listen, class, this is how everything works, fo SHO! "
+  #   response += "*drops flat-out insane knowledge bomb* "
+  #   response += "... You're welcome. *saunters away*"
+  #   response
+  # end
 
   def salary=(new_salary)
     puts "This better be good!"
